修改EGLBase-->EGLCore
